Thread I2CP lookups
until we have a nonblocking I2CP lookup API.
Most of the time, this won't be needed, as our cache
will have it, but if additional announces come in over
the lifetime, we'll have to do a I2CP lookup.
Even then, the I2CP cache may have it,
or the router should have the LS, so it should be quick.